refactor(about-us): extract VideoPlayer component to shared UI components
